# openapi.yaml
openapi: 3.1.0
info:
  title: Marconi Inmobiliaria Analytics & Geocoding API
  description: |
    Comprehensive API for real estate analytics tracking and geocoding services.
    Includes GDPR-compliant user session management, property view tracking,
    lead generation attribution, and external geocoding proxy services.
  version: "1.0.0"
  contact:
    name: Marconi Inmobiliaria API Support
    email: contact@marconi-inmobiliaria.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://marconi-inmobiliaria.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - bearerAuth: []
  - apiKey: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    StandardResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Operation success status
        error:
          type: string
          description: Error message if success is false
        message:
          type: string
          description: Success message

    # ==== GEOCODING SCHEMAS ====
...
    CreateInteractionRequest:
      type: object
      required:
        - session_id
        - event_type
      properties:
        session_id:
          type: string
          format: uuid
          description: Analytics session ID
        property_id:
          type: integer
          minimum: 1
          description: Property ID if interaction is property-specific
        event_type:
          type: string
          enum: [
            click, scroll, form_field_focus, form_submit, page_load,
            page_unload, contact_click, phone_click, whatsapp_click,
            email_click, image_view, gallery_interaction
          ]
          description: Type of user interaction
        event_target:
          type: string
          maxLength: 255
          description: CSS selector or element identifier
        page_url:
          type: string
          format: uri
          description: URL where interaction occurred
        event_data:
          type: object
          description: Additional event-specific data
          additionalProperties: true
        timestamp:
          type: integer
          description: Unix timestamp of interaction (optional, defaults to current time)
    BatchInteractionRequest:
      type: object
      required:
        - interactions
      properties:
        interactions:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/CreateInteractionRequest'

    # ==== LEAD GENERATION SCHEMAS ====
...
    DashboardFilters:
      type: object
      properties:
        start_date:
          type: string
          format: date
          description: Start date for analytics period (ISO 8601)
          example: "2024-01-01"
        end_date:
          type: string
          format: date
          description: End date for analytics period (ISO 8601)
          example: "2024-01-31"
        property_ids:
          type: array
          items:
            type: integer
          description: Filter by specific property IDs
        lead_source_ids:
          type: array
          items:
            type: integer
          description: Filter by specific lead source IDs
        utm_source:
          type: string
          description: Filter by UTM source
        utm_campaign:
          type: string
          description: Filter by UTM campaign
        device_type:
          type: string
          enum: [desktop, mobile, tablet]
          description: Filter by device type
        country_code:
          type: string
          pattern: '^[A-Z]{2}$'
          description: Filter by country code
    DashboardStatsResponse:
      allOf:
        - $ref: '#/components/schemas/StandardResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                total_sessions:
                  type: integer
                  description: Total number of analytics sessions
                total_property_views:
                  type: integer
                  description: Total property view events
                unique_property_views:
                  type: integer
                  description: Unique property views (deduplicated)
                total_leads:
                  type: integer
                  description: Total leads generated
                conversion_rate:
                  type: number
                  format: float
                  description: Overall conversion rate (leads/sessions)
                avg_time_on_page:
                  type: number
                  format: float
                  description: Average time spent on property pages (seconds)
                top_properties:
                  type: array
                  items:
                    type: object
                    properties:
                      property_id:
                        type: integer
                      title:
                        type: string
                      metric_value:
                        type: number
                      unique_views:
                        type: integer
                      leads:
                        type: integer
                top_lead_sources:
                  type: array
                  items:
                    type: object
                    properties:
                      source_id:
                        type: integer
                      source_name:
                        type: string
                      leads_count:
                        type: integer
                      conversion_rate:
                        type: number
                traffic_by_device:
                  type: array
                  items:
                    type: object
                    properties:
                      device_type:
                        type: string
                      sessions:
                        type: integer
                      percentage:
                        type: number
                daily_stats:
                  type: array
                  items:
                    type: object
                    properties:
                      date:
                        type: string
                        format: date
                      sessions:
                        type: integer
                      views:
                        type: integer
                      leads:
                        type: integer
            filters:
              $ref: '#/components/schemas/DashboardFilters'

    # ==== ERROR SCHEMAS ====
...
paths:
...
  /analytics/interaction:
    post:
      tags:
        - Analytics
      summary: Record user interaction event
      description: |
        Records a single user interaction event (click, scroll, form interaction, etc.)
        for UX analysis and user behavior tracking.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInteractionRequest'
      responses:
        '200':
          description: Interaction recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                success: true
                message: "Interaction recorded successfully"
        '400':
          description: Validation error - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'

    put:
      tags:
        - Analytics
      summary: Record multiple interaction events (batch)
      description: |
        Records multiple user interaction events in a single request.
        Limited to 100 interactions per batch for performance.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchInteractionRequest'
      responses:
        '200':
          description: Batch interactions recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
              example:
                success: true
                message: "5 interactions recorded successfully"
        '400':
          description: Validation error - invalid batch data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'

  # ==== LEAD GENERATION ENDPOINTS ====
...
  /analytics/dashboard:
    get:
      tags:
        - Analytics
      summary: Get dashboard analytics (basic)
      description: |
        Retrieves comprehensive analytics data for the admin dashboard
        with optional filtering. Defaults to last 30 days if no date range provided.
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
          description: Start date (ISO 8601)
          example: "2024-01-01"
        - name: end_date
          in: query
          schema:
            type: string
            format: date
          description: End date (ISO 8601)
          example: "2024-01-31"
        - name: property_ids
          in: query
          schema:
            type: string
          description: Comma-separated property IDs
          example: "1,2,3"
        - name: lead_source_ids
          in: query
          schema:
            type: string
          description: Comma-separated lead source IDs
          example: "1,2,3"
        - name: utm_source
          in: query
          schema:
            type: string
          description: Filter by UTM source
        - name: utm_campaign
          in: query
          schema:
            type: string
          description: Filter by UTM campaign
        - name: device_type
          in: query
          schema:
            type: string
            enum: [desktop, mobile, tablet]
          description: Filter by device type
        - name: country_code
          in: query
          schema:
            type: string
            pattern: '^[A-Z]{2}$'
          description: Filter by country code
      responses:
        '200':
          description: Dashboard analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStatsResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'

    post:
      tags:
        - Analytics
      summary: Get advanced dashboard analytics
      description: |
        Retrieves specific analytics metrics based on requested metric types.
        Supports complex filtering and multiple metric sets in a single request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/DashboardFilters'
                - type: object
                  properties:
                    metrics:
                      type: array
                      items:
                        type: string
                        enum: [overview, properties, campaigns, sources, devices, all]
                      default: [overview]
                      description: Specific metrics to retrieve
                      example: ["overview", "properties", "campaigns"]
      responses:
        '200':
          description: Advanced analytics retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        description: Requested analytics data grouped by metric type
                        additionalProperties: true
                      filters:
                        $ref: '#/components/schemas/DashboardFilters'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardResponse'

  # ==== GDPR COMPLIANCE ENDPOINTS ====
...